refactor: MVC 패턴 적용
InputView에서 실행하던 OutputView를 controller로 이동
format과 관련된 기능 Util로 이동
