+++ src/main/java/lotto/view/InputView.java
@@ -8,35 +8,12 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class InputView {
-    public static String inputTotalBudget() {
-        OutputView.printCashInput();
-        return getInput();
-    }
 
-    public static List<Integer> inputWinningNumbers() {
-        OutputView.printWinningNumberInput();
-        return formatInput(getInput());
-    }
-
-    public static String inputBonusNumber() {
-        OutputView.printBonusNumberInput();
-        return getInput();
-    }
-
-    private static String getInput() {
+    public static String getInput() {
         return Console.readLine();
     }
 
-    private static List<Integer> formatInput(String input) {
-        return convertStringListToIntList(separateStringByComma(removeSpace(input)));
-    }
 
-    private static List<String> separateStringByComma(String input) {
-        return Arrays.asList(input.split(","));
-    }
 
-    private static List<Integer> convertStringListToIntList(List<String> list) {
-        return list.stream().map(s -> Integer.parseInt(s))
-                .collect(Collectors.toList());
-    }
+
 }
